+++ cmd/block/info/output.go
@@ -51,6 +51,7 @@ func output(ctx context.Context, data *dataOut) (string, error) {
 func outputBlockGeneral(ctx context.Context,
 	verbose bool,
 	slot phase0.Slot,
+	blockRoot phase0.Root,
 	bodyRoot phase0.Root,
 	parentRoot phase0.Root,
 	stateRoot phase0.Root,
@@ -67,8 +68,9 @@ func outputBlockGeneral(ctx context.Context,
 	res.WriteString(fmt.Sprintf("Slot: %d\n", slot))
 	res.WriteString(fmt.Sprintf("Epoch: %d\n", phase0.Epoch(uint64(slot)/slotsPerEpoch)))
 	res.WriteString(fmt.Sprintf("Timestamp: %v\n", time.Unix(genesisTime.Unix()+int64(slot)*int64(slotDuration.Seconds()), 0)))
-	res.WriteString(fmt.Sprintf("Block root: %#x\n", bodyRoot))
+	res.WriteString(fmt.Sprintf("Block root: %#x\n", blockRoot))
 	if verbose {
+		res.WriteString(fmt.Sprintf("Body root: %#x\n", bodyRoot))
 		res.WriteString(fmt.Sprintf("Parent root: %#x\n", parentRoot))
 		res.WriteString(fmt.Sprintf("State root: %#x\n", stateRoot))
 	}
@@ -271,13 +273,19 @@ func outputAltairBlockText(ctx context.Context, data *dataOut, signedBlock *alta
 	res := strings.Builder{}
 
 	// General info.
+	blockRoot, err := signedBlock.Message.HashTreeRoot()
+	if err != nil {
+		return "", errors.Wrap(err, "failed to obtain block root")
+	}
 	bodyRoot, err := signedBlock.Message.Body.HashTreeRoot()
 	if err != nil {
-		return "", errors.Wrap(err, "failed to generate block root")
+		return "", errors.Wrap(err, "failed to generate body root")
 	}
+
 	tmp, err := outputBlockGeneral(ctx,
 		data.verbose,
 		signedBlock.Message.Slot,
+		blockRoot,
 		bodyRoot,
 		signedBlock.Message.ParentRoot,
 		signedBlock.Message.StateRoot,
@@ -349,6 +357,10 @@ func outputPhase0BlockText(ctx context.Context, data *dataOut, signedBlock *phas
 	res := strings.Builder{}
 
 	// General info.
+	blockRoot, err := signedBlock.Message.HashTreeRoot()
+	if err != nil {
+		return "", errors.Wrap(err, "failed to obtain block root")
+	}
 	bodyRoot, err := signedBlock.Message.Body.HashTreeRoot()
 	if err != nil {
 		return "", errors.Wrap(err, "failed to generate block root")
@@ -356,6 +368,7 @@ func outputPhase0BlockText(ctx context.Context, data *dataOut, signedBlock *phas
 	tmp, err := outputBlockGeneral(ctx,
 		data.verbose,
 		signedBlock.Message.Slot,
+		blockRoot,
 		bodyRoot,
 		signedBlock.Message.ParentRoot,
 		signedBlock.Message.StateRoot,