+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+dev:
+  - show both block and body root in "block info"
+
 1.13.0:
   - rework and provide additional information to "chain status" output